tweet_consumer/requirements.txt
kafka-python
py7zr
nltk
scikit-learn